ml-backend/requirements.txt
flask
flask-cors
kafka-python
transformers
tensorflow